Fix pointer truncation

+++ include/sf33rd/Source/PS2/mc/mcsub.h
@@ -5,10 +5,10 @@
 
 typedef struct _file {
     // total size: 0x10
-    s32 flag; // offset 0x0, size 0x4
-    s8 *name; // offset 0x4, size 0x4
-    s32 bufs; // offset 0x8, size 0x4
-    s32 size; // offset 0xC, size 0x4
+    s32 flag;      // offset 0x0, size 0x4
+    s8 *name;      // offset 0x4, size 0x4
+    intptr_t bufs; // offset 0x8, size 0x4
+    s32 size;      // offset 0xC, size 0x4
 };
 
 typedef struct _memcard_file {
